requirements.txt
# Core dependencies
click>=8.0.0
pydantic>=2.0.0
pydantic-settings>=0.1.0
python-dotenv>=1.0.0
rich>=13.0.0
taskw>=2.0.0

# GitHub integration
PyGithub>=1.59.0

# Voice synthesis
pyttsx3>=2.90
gTTS>=2.3.0
pygame>=2.5.0

# Image generation for Kanban
Pillow>=10.0.0
matplotlib>=3.7.0

# Testing dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-asyncio>=0.21.0
coverage>=7.0.0

google-generativeai>=0.3.0
openai>=1.0.0